requirements.txt
# Core numerical and scientific libraries
numpy>=1.21.0
scipy>=1.7.0

# Machine learning utilities
scikit-learn>=1.0.2

# FAISS for similarity search (CPU version)
faiss-cpu>=1.7.2

# Data handling
pandas>=1.3.0

# Web framework (if your app.py uses Flask)
Flask>=2.0.2

